+++ src/richie/apps/search/viewsets/courses.py
@@ -7,6 +7,7 @@ from elasticsearch.exceptions import NotFoundError
 from rest_framework.response import Response
 from rest_framework.viewsets import ViewSet
 
+from ..defaults import ES_PAGE_SIZE
 from ..filter_definitions import FILTERS
 from ..indexers import ES_INDICES
 from ..utils.viewsets import AutocompleteMixin, ViewSetMetadata
@@ -56,7 +57,7 @@ class CoursesViewSet(AutocompleteMixin, ViewSet):
             body=body,
             # Directly pass meta-params through as arguments to the ES client
             from_=offset,
-            size=limit or settings.ES_DEFAULT_PAGE_SIZE,
+            size=limit or getattr(settings, "RICHIE_ES_PAGE_SIZE", ES_PAGE_SIZE),
         )
 
         response_object = {