+++ tests/apps/search/test_autocomplete_courses.py
@@ -103,11 +103,7 @@ class AutocompleteCoursesTestCase(TestCase):
             }
             for course in courses
         ]
-        bulk(
-            actions=actions,
-            chunk_size=settings.ES_CHUNK_SIZE,
-            client=settings.ES_CLIENT,
-        )
+        bulk(actions=actions, chunk_size=500, client=settings.ES_CLIENT)
         indices_client.refresh()
 
         response = self.client.get(f"/api/v1.0/courses/autocomplete/?{querystring:s}")