+++ tests/apps/search/test_query_courses.py
@@ -291,11 +291,7 @@ class CourseRunsCoursesQueryTestCase(TestCase):
             }
             for course_id, course_run_ids in courses_definition
         ]
-        bulk(
-            actions=actions,
-            chunk_size=settings.ES_CHUNK_SIZE,
-            client=settings.ES_CLIENT,
-        )
+        bulk(actions=actions, chunk_size=500, client=settings.ES_CLIENT)
         indices_client.refresh()
 
         response = self.client.get(f"/api/v1.0/courses/?{querystring:s}")