+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## Changed
+
+- Simplify search settings and provide defaults for those unlikely to be
+  customized.
+
 ## [1.0.0-beta.5] - 2019-04-11
 
 ### Added